On submit function and some changes in styles
